for_bisenet/requirements.txt
paddlepaddle
heapq
traitlets
opencv-python
numpy
torch
git+https://github.com/WenmuZhou/reprod_log.git --force-reinstall